moved some interfaces into namespace
