+++ Core/Menu/IGMDataItem/I_Color.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 
-namespace OpenVIII
+namespace OpenVIII.IGMDataItem
 {
     public interface I_Color
     {
